TS 폐지 및 API 통폐합
