+++ views/main_css/js/route/topic.js
@@ -1,11 +1,14 @@
 function opennamu_do_remove_blind_thread() {
-    var style = document.querySelector('#opennamu_remove_blind');
-    if (style !== null) {
-        if (style.innerHTML !== "") {
+    const style = document.querySelector('#opennamu_remove_blind');
+    if(style !== null) {
+        if(style.innerHTML !== "") {
             style.innerHTML = '';
-        }
-        else {
-            style.innerHTML = "\n                .opennamu_comment_blind_js {\n                    display: none;\n                }\n            ";
+        } else {
+            style.innerHTML = `
+                .opennamu_comment_blind_js {
+                    display: none;
+                }
+            `;
         }
     }
 }
